+++ src/frontend/apps/lti_site/apps/classroom/types/ClassroomAppData.ts
@@ -1,0 +1,9 @@
+import { Nullable } from 'lib-common';
+
+import { AppData } from 'types/AppData';
+import { Classroom, modelName } from './models';
+
+export interface ClassroomAppData extends Omit<AppData, 'modelName'> {
+  modelName: modelName.CLASSROOMS;
+  classroom?: Nullable<Classroom>;
+}